Minor fixes

+++ mrp2_display/src/display.cpp
@@ -177,7 +177,7 @@ int main(int argc, char **argv)
 	}
 
   	bumpers_sub = n.subscribe<std_msgs::Int32MultiArray>("bumpers", 1, &bumpersCallback);
-  	battery_volt_sub = n.subscribe<std_msgs::Int32>("/hw_monitor/batt_volt", 1, &batteryVoltCallback);
+  	battery_volt_sub = n.subscribe<std_msgs::Float32>("/hw_monitor/batt_volt", 1, &batteryVoltCallback);
 
   	inputs_pub = n.advertise<std_msgs::Int32MultiArray>("panel_inputs", 1);
   	ros::ServiceServer service = n.advertiseService("/panel_outputs/gpio", gpio);
